+++ src/burst2safe/auth.py
@@ -6,6 +6,7 @@ from typing import Tuple
 
 
 EARTHDATA_HOST = 'urs.earthdata.nasa.gov'
+TOKEN_ENV_VAR = 'EARTHDATA_TOKEN'
 
 
 def get_netrc() -> Path:
@@ -57,21 +58,51 @@ def find_creds_in_netrc(service) -> Tuple[str, str]:
     return None, None
 
 
-def get_earthdata_credentials() -> Tuple[str, str]:
-    """Get NASA EarthData credentials from the environment or netrc file.
+def write_credentials_to_netrc_file(username: str, password: str) -> None:
+    """Write credentials to netrc file
+
+    Args:
+        username: NASA EarthData username
+        password: NASA EarthData password
+    """
+    netrc_file = get_netrc()
+    if not netrc_file.exists():
+        netrc_file.touch()
+
+    with open(netrc_file, 'a') as f:
+        f.write(f'machine {EARTHDATA_HOST} login {username} password {password}\n')
+
+
+def check_earthdata_credentials(append=False) -> str:
+    """Check for NASA EarthData credentials in the netrc file or environment variables.
+    Will preferentially use the netrc file, and write credentials to the netrc file if found in the environment.
+
+    Args:
+        append: Whether to append the credentials to the netrc file if creds found in the environment
 
     Returns:
-        Tuple of the NASA EarthData username and password
+        The location of the preferred credentials ('netrc' or 'token')
     """
-    username, password = find_creds_in_env('EARTHDATA_USERNAME', 'EARTHDATA_PASSWORD')
-    if username and password:
-        return username, password
+    if os.getenv(TOKEN_ENV_VAR):
+        return 'token'
 
     username, password = find_creds_in_netrc(EARTHDATA_HOST)
     if username and password:
-        return username, password
+        return 'netrc'
+
+    username, password = find_creds_in_env('EARTHDATA_USERNAME', 'EARTHDATA_PASSWORD')
+    if username and password:
+        if append:
+            write_credentials_to_netrc_file(username, password)
+            return 'netrc'
+        else:
+            raise ValueError(
+                'NASA Earthdata credentials only found in environment variables,'
+                'but appending to netrc file not allowed. Please allow appending to netrc.'
+            )
 
     raise ValueError(
-        'Please provide NASA EarthData credentials via the '
-        'EARTHDATA_USERNAME and EARTHDATA_PASSWORD environment variables, or your netrc file.'
+        'Please provide NASA Earthdata credentials via your .netrc file,'
+        'the EARTHDATA_USERNAME and EARTHDATA_PASSWORD environment variables,'
+        'or an EDL Token via the EARTHDATA_TOKEN environment variable.'
     )