+++ src/burst2safe/burst2safe.py
@@ -8,8 +8,9 @@ from typing import Optional
 from shapely.geometry import Polygon
 
 from burst2safe import utils
+from burst2safe.download import download_bursts
 from burst2safe.safe import Safe
-from burst2safe.search import download_bursts, find_bursts
+from burst2safe.search import find_bursts
 
 
 DESCRIPTION = """Convert a set of ASF burst SLCs to the ESA SAFE format.